+++ cmd/frontend/graphqlbackend/external_services.go
@@ -3,6 +3,7 @@ package graphqlbackend
 import (
 	"context"
 	"fmt"
+	"os"
 	"strings"
 	"sync"
 
@@ -17,6 +18,10 @@ import (
 	"github.com/sourcegraph/sourcegraph/pkg/repoupdater"
 )
 
+func haveExternalServiceConfigFile() bool {
+	return os.Getenv("EXTSVC_CONFIG_FILE") != ""
+}
+
 func (r *schemaResolver) AddExternalService(ctx context.Context, args *struct {
 	Input *struct {
 		Kind        string
@@ -27,6 +32,9 @@ func (r *schemaResolver) AddExternalService(ctx context.Context, args *struct {
 	// 🚨 SECURITY: Only site admins may add external services.
 	if err := backend.CheckCurrentUserIsSiteAdmin(ctx); err != nil {
 		return nil, err
+	}
+	if haveExternalServiceConfigFile() {
+		return nil, errors.New("adding external service not allowed when using EXTSVC_CONFIG_FILE")
 	}
 
 	externalService := &types.ExternalService{
@@ -61,6 +69,9 @@ func (*schemaResolver) UpdateExternalService(ctx context.Context, args *struct {
 	// 🚨 SECURITY: Only site admins are allowed to update the user.
 	if err := backend.CheckCurrentUserIsSiteAdmin(ctx); err != nil {
 		return nil, err
+	}
+	if haveExternalServiceConfigFile() {
+		return nil, errors.New("updating external service not allowed when using EXTSVC_CONFIG_FILE")
 	}
 
 	if args.Input.Config != nil && strings.TrimSpace(*args.Input.Config) == "" {
@@ -112,6 +123,9 @@ func (*schemaResolver) DeleteExternalService(ctx context.Context, args *struct {
 	// 🚨 SECURITY: Only site admins can delete external services.
 	if err := backend.CheckCurrentUserIsSiteAdmin(ctx); err != nil {
 		return nil, err
+	}
+	if haveExternalServiceConfigFile() {
+		return nil, errors.New("deleting external service not allowed when using EXTSVC_CONFIG_FILE")
 	}
 
 	id, err := unmarshalExternalServiceID(args.ExternalService)